# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Skymark Eatery API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /specials/today:
    get:
      operationId: getTodaySpecial
      tags: [specials]
      summary: Get today's featured special
      responses:
        "200":
          description: Today's special
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Special"
        "404":
          description: No special today
  /specials:
    get:
      operationId: listSpecials
      tags: [specials]
      summary: List all specials
      responses:
        "200":
          description: List of specials
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Special"
    post:
      operationId: createSpecial
      tags: [specials]
      summary: Create or update today's special
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSpecialBody"
      responses:
        "201":
          description: Special created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Special"
  /specials/{id}:
    patch:
      operationId: updateSpecial
      tags: [specials]
      summary: Update a special
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSpecialBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Special"

  # ─── ORDERS ──────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Special:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        price:
          type: ["number", "null"]
        imageUrl:
          type: ["string", "null"]
        activeDate:
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
      required: [id, title, description, activeDate, isActive, createdAt]
    CreateSpecialBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        price:
          type: ["number", "null"]
        imageUrl:
          type: ["string", "null"]
        activeDate:
          type: string
        isActive:
          type: boolean
      required: [title, description, activeDate]
    UpdateSpecialBody:
      type: object
      properties:
        title:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        price:
          type: ["number", "null"]
        imageUrl:
          type: ["string", "null"]
        activeDate:
          type: ["string", "null"]
        isActive:
          type: ["boolean", "null"]

    # ─── ORDERS ────────────────────────────────────────────────────────────────
...